Corrected num_events to store in lm_to_proj_data

+++ src/include/stir/listmode/LmToProjData.h
@@ -242,7 +242,7 @@ protected:
   bool use_tof;
   int num_segments_in_memory;
   // TODO make long (or even unsigned long) but can't do this yet because we can't parse longs yet
-  unsigned long int num_events_to_store;
+  int num_events_to_store;
   int max_segment_num_to_process;
 
   //! Toggle readable output on stdout or actual projdata
